fix: update pivot expr on to first param (#252)
The pivot function now has `on` as the first param instead of `index`,
as noted in the [python 1.0 upgrade
documentation](https://docs.pola.rs/releases/upgrade/1/#update-resulting-column-names-in-pivot-when-pivoting-by-multiple-values):
> columns has been renamed to on, and is now the first positional
argument.
<img width="1099" alt="image"
src="https://github.com/user-attachments/assets/1110a76d-2212-4fd0-9508-ce3bd3b619ba">

Added some test coverage from examples in docs to capture this. Existing
test coverage was using the same column as `index` and `on`, resulting
in this not being caught.